// modules/transfer/modules/transfer-ach/enum/ach-transfer-control-name.enum.ts
export enum AttributeFormTransferAch {
  ACCOUNT_CREDIT_NAME = 'accountCreditName',
  ACCOUNT_CREDIT_CURRENCY = 'accountCreditCurrency',
  ACCOUNT_CREDIT_TYPE = 'accountCreditType',
  ACCOUNT_CREDIT_BANK = 'accountCreditBank',
  ACCOUNT_CREDIT_ALIAS = 'accountCreditAlias',
  ACCOUNT_DEBITED = 'accountDebited',
  AMOUNT = 'amount',
  COMMENT = 'comment',
  NAME_ACCOUNT = 'nameAccount',
  IDENTIFY_BENEFICIARY = 'identifyBeneficiary',
  ALIAS = 'alias',
  SCHEDULE = 'schedule',
  DATE = 'date',
  HOUR = 'hour',
  FILTER_SEARCH = 'nameAlias',
  SELECT_TYPE_FILTER = 'typeFilter',
}